requirements.txt
langgraph
langchain-google-genai
python-dotenv
langchain
numpy
pandas